[kf5bluezqt] Remove two patches. JB#61506
Removed
- "Don't connect to signals with QVariantMapMap parameters. Fixes JB#36800"
- "Add Manager::monitorObjectManagerInterfaces. Contributes to JB#40602"

With the first patch, pairing a bluetooth LE device with pin ended
up process stuck. It was introduced to avoid some problem for handling
receiving file over bluetooth, but that keeps on working even without.
Generally the upstream signal handling code _should_ work already so
the patch was somewhat fishy.

The latter patch built a bit upon the earlier, but the custom property
was used only on one short-lived process. Thus not seeing reason for
keeping it. Less delta to upstream is good.